Move pkg/kubectl/util/resource to staging
