+++ staging/src/k8s.io/kubectl/pkg/util/resource/resource.go
